adding nesting to storybook tree
No Story

- rearranged directory structure to better suit our needs
- added core and rent top level nesting
