+++ stories/core/Collapsible.js
