+++ stories/core/FilterPanel.js
