+++ stories/core/Modal.js
@@ -3,6 +3,7 @@ import { action } from '@storybook/addon-actions'
 import { boolean } from '@storybook/addon-knobs'
 import { Modal } from 'react-ui-core/src'
 import { Button } from 'react-ui-core/src/Button'
+import StorybookTheme from '../theme/Storybook.css'
 
 const CloseButton = props => (
   <Button {...props}>
@@ -42,13 +43,17 @@ export const ModalNoOverlayClose = (
 export const ModalPopup = () => {
   const isOpen = boolean('isOpen', false)
   return [
-    <div>
+    <div
+      className={StorybookTheme.Story_ModalClickText}
+      key="activate-modal"
+    >
       Activate through knobs to open
     </div>,
     <Modal
       isOpen={isOpen}
+      key="modal"
     >
-      <div>Modal body</div>
+      <div>{"I'm open! Woohoo!"}</div>
     </Modal>,
   ]
 }
