+++ stories/core/ResponsiveTemplate.js
