+++ stories/core/SampleRatingItem/Circle.js
