+++ stories/core/SampleRatingItem/Square.js
