+++ stories/core/SampleRatingItem/index.js
