+++ stories/core/Text.js
