--- stories/examples/index.js
@@ -1,95 +1,0 @@
-export {
-  default as CompleteForm,
-} from './CompleteForm'
-
-export {
-  default as FilterPanel,
-} from './FilterPanel'
-
-export {
-  default as RangeSlider,
-} from './RangeSlider'
-
-export {
-  default as Grid,
-} from './Grid'
-
-export {
-  default as ResponsiveTemplate,
-} from './ResponsiveTemplate'
-
-export {
-  DefaultButton,
-  ButtonWithColor,
-  ButtonLarge,
-  ButtonSmall,
-} from './Button'
-
-export {
-  default as Text,
-} from './Text'
-
-export {
-  DefaultField,
-  SelectField,
-  CheckboxField,
-  RadioField,
-} from './Field'
-
-export {
-  DefaultModal,
-  CloseModal,
-  ModalNoOverlayClose,
-  ModalPopup,
-} from './Modal'
-
-export {
-  MapExample,
-} from './Map'
-
-export {
-  Collapse,
-} from './Collapsible'
-
-export {
-  DefaultCounter,
-  CounterStep,
-  CounterText,
-  CounterCustomButtons,
-} from './Counter'
-
-export {
-  DefaultRatings,
-  SquareRatings,
-  CircleRatings,
-  ThreeRatings,
-  ManyRatings,
-  Partial,
-  PartialTwoColor,
-} from './Ratings'
-
-export {
-  DefaultList,
-  HorizontalList,
-  ListWithPassedNodeTypes,
-  ListWithOwnItemComponent,
-} from './List'
-
-export {
-  DefaultLeadModal,
-} from './rent'
-
-export {
-  DefaultCard,
-} from './Card'
-
-export {
-  DefaultRadioGroup,
-  ButtonRadioGroup,
-  VerticalRadioGroup,
-} from './RadioGroup'
-
-export {
-  DefaultTitle,
-  EmbeddedLinkWithinTitle,
-} from './Title'
