+++ stories/theme/Storybook.css
@@ -5,3 +5,12 @@
 .Story-alignCenter {
   text-align: center;
 }
+
+.Story_ModalClickText {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  text-transform: uppercase;
+  color: #ccc;
+}
