+++ stories/theme/core/Modal.css
@@ -28,12 +28,3 @@
   align-items: center;
   justify-content: center;
 }
-
-.ClickText {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  text-transform: uppercase;
-  color: #ccc;
-}
