+++ stories/theme/core/RadioGroup.css
@@ -1,5 +1,4 @@
 .RadioGroup {
-  background: rgb(207, 207, 207);
   border-radius: 5px;
   padding: .5rem 1rem;
 }
