+++ stories/theme/map/index.js
@@ -1,0 +1,6 @@
+import { compose } from 'react-themed'
+import Mapbox from './Map.css'
+
+export default compose({},
+  Mapbox,
+)
